Address the #389 review: pin the wide-source test, dedupe asCount
Five review comments.

The wide-source regression test could silently stop testing anything. The only
thing making it a wide-source test was an em-dash inside the compiled string,
so if that ever stopped round-tripping -- or the fixture's compile:in: (which
resumes CompileWarning with nil) swallowed a compile problem -- the method
would be plain ASCII, the shim would never be exercised, and the test would
still pass. It now reads the STORED source back and asserts the em-dash is
present before asserting anything else.

Its outcome assertions were also weak. `deny: analysisJsonString includesSubstring:
'does not parse'` is implied by `decline isNil`, so it could not fail on its own.
Replaced with the positive assertions the ASCII sibling test uses -- the ivar edit
and the method recompile are both staged -- plus the failure mode the wide path
most needs covered and previously passed regardless: that the em-dash survives
into the RECOMPILED source rather than coming back mangled or truncated.

Verified those assertions have teeth rather than passing incidentally. Probed on
both boundaries: the stored source really is a DoubleByteString, the em-dash is
genuinely present, and it is still present after the rewrite.

`wide` renamed to `wideSource` (it holds the source, not the character), with the
em-dash hoisted into its own `emDash` so it is named once and reused.

asCount now truncates a fractional count toward zero. The doc enumerated what
clamps to 0, which implied everything else was a usable count -- but 2.5 passed
straight through into `total` / `applied` / `nextOffset` and the pagination
arithmetic, where half a change cannot render and a fractional offset asks for a
page boundary that does not exist.

And the helper is deduped. There were THIRTEEN byte-identical private copies
across client/src/refactoring/*Preview.ts (verified identical before merging;
only methodRelocationPreview's carried `export`), so a reader landing in any of
the other twelve met an undocumented helper and had to re-derive the rules. One
canonical previewCounts.ts now, imported by all thirteen including
methodRelocationPreview, so no privileged copy remains. Its tests moved out of
methodRelocation.test.ts to sit with the code, expanded with null / {} /
-Infinity and the truncation cases.

Verified: RB SUnit 527/527 in-stone on BOTH 3.6.2 and 3.7.5; client suite 4950
(3.6.2) / 4952 (3.7.5); lint, format and compile clean.
